requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
